move check to ForbiddenPropertyRemovalAnalyzer service
